README.md: change the build instruction
After go1.16, `go get` runs in modules mode, and it's not the
recommended way of installing binaries nor a way to fetch source
code. Given that this example repo is designed for new go users
who want to learn, it is better to guide users to clone the repo
using the git command.

Updated the instruction for appengine-hello since goapp is gone.
I didn't test the appengine-hello instruction myself and not sure
if it still runs.

Change-Id: I74a829203def9b062bee75637c7d2c6cef448a2f