ragserver: don't try to build on 32-bit Go ports
Trying to only build some of these commands on 32-bit ports fails
due to a build error in a dependency:

	$ GOOS=linux GOARCH=386 go build golang.org/x/example/ragserver/ragserver
	# github.com/weaviate/weaviate/entities/vectorindex/common
	$GOMODCACHE/github.com/weaviate/weaviate@v1.26.1/entities/vectorindex/common/config.go:59:14: constant 9223372036854775807 overflows int

That particular error is due to it trying to convert math.MaxInt64
to an 'int' type.

Working on 32-bit Go ports seems out of scope for these examples.
Add build constraints that make this limitation explicit.

Runs into golang/go#52287 (unfortunately).

Change-Id: I9db291fbb06cfd9eed1df318aff6a259b7603f95