example/wave: update TOC anchors to remove **, _, and `.
When GitHub's Markdown parser creates anchors for sections, it removes
formatting directives like **, _, and `. See the discussion at
https://support.github.com/ticket/personal/0/3324945

The links to the sections with formatting directives in the current TOC at
https://github.com/golang/example/blob/master/slog-handler-guide/README.md
are broken for this reason.

This change modifies weave to also remove those characters from the anchor
links it generates.

This change also replaces the deprecated os.SET_SEEK with io.SeekStart and
removes two unused functions.

Change-Id: I66e3aa8140e14146bdb349c4ccfb773fe7414c67